DOCS-6679 update connection/install screenshots
Re-took screenshots of the connections dialog and the OS X installation dialog,
which should be cleaner and free of artifacts.

The images *are* on the larger side as they include the OS X native window drop
shadow. A future commit can take a pass at this and trim the excess with a
professional image editor.

Signed-off-by: kay <kay.kim@10gen.com>